+++ R/7-entropy.R
@@ -26,11 +26,12 @@ objective.keepEntropy <- function(codon, ARGS){
 
   W <- do.call(ARGS$weights, list(codon, n.amalgams = ARGS$n.amalgams))
   A <- ARGS$x %*% W
+  A <- sweep(A, 1, rowSums(A), "/")
 
-  # Don't allow zeros!
-  if(any(A == 0)){
-    return(-1000000)
-  }
+  # # Don't allow zeros!
+  # if(any(A == 0)){
+  #   return(-1000000)
+  # }
 
   if(ARGS$asSLR){
     stop("This objective does not support summed log-ratios.")
@@ -54,11 +55,12 @@ objective.diffEntropy <- function(codon, ARGS){
 
   W <- do.call(ARGS$weights, list(codon, n.amalgams = ARGS$n.amalgams))
   A <- ARGS$x %*% W
+  A <- sweep(A, 1, rowSums(A), "/")
 
-  # Don't allow zeros!
-  if(any(A == 0)){
-    return(-1000000)
-  }
+  # # Don't allow zeros!
+  # if(any(A == 0)){
+  #   return(-1000000)
+  # }
 
   if(ARGS$asSLR){
     stop("This objective does not support summed log-ratios.")